backend/requirements.txt
fastapi
uvicorn
qdrant-client
requests
python-dotenv
pydantic